Fixed detection of imported class
This fixes the bug where another (incorrect) FQN can be wrong detected if the class name ended with a suffix that is identical to the short class name being searched for.
For example: if the script is searching for the imported FQN for `FactoryInterface`, the `use` statement for a `ContainerFactoryInterface` may be detected instead.
This is only the case when the incorrect and longer FQN appears before the correct one.